backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.9
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
django-timezone-field==4.2.3
djangorestframework==3.12.4
kombu==5.4.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
redis==5.0.8
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
